rework PlayIntroTitle.
Things that happen in a known order have been taken out of the loop.
A class was written to help with prevent copy-pasting code blocks.
The class contains variables and methods for the things that happen outside the original loop.
The original loop now only contains things that depend on user input.